% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{This function, given two numerical input, return some basic statistical operation}
\usage{
analysis(x, y)
}
\arguments{
\item{x}{set of numerical input}

\item{y}{set of numerical input}
}
\value{
mean, variance, standard deviation, maximum value and its position, minimum value
and its position, correlation, simple regression, line plot, bar plot, box plot and scatter plot.
}
\description{
This function, given two numerical input, return some basic statistical operation
}
\examples{
analysis(data$y,data$Italy)
}
