% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{data: this function allowa the use to import the dataset used in this package}
\usage{
data()
}
\value{
a dataset called "data"
}
\description{
data: this function allowa the use to import the dataset used in this package
}
\examples{
data()
}
