% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{This function allows the user to view a dataset of total population of Italy, Germany and France
during the period from 1970 to 2018. In order to obtain this dataset into your environment
type the following command: data <- data.frame(getDataset())}
\usage{
getDataset()
}
\value{
dataset called "data"
}
\description{
This function allows the user to view a dataset of total population of Italy, Germany and France
during the period from 1970 to 2018. In order to obtain this dataset into your environment
type the following command: data <- data.frame(getDataset())
}
\examples{
getDataset()
}
