% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{importdata}
\alias{importdata}
\title{importdata: this function allowa the use to import the dataset used in this package}
\usage{
importdata()
}
\value{
a dataset called "data"
}
\description{
importdata: this function allowa the use to import the dataset used in this package
}
\examples{
importdata()
}
