% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{PreAn}
\alias{PreAn}
\title{ShinyApp: this function allows the user to create some interactive grpahs with shiny and plotly}
\usage{
PreAn()
}
\value{
Line plot, bar plot, box plot, scatter plot
}
\description{
ShinyApp: this function allows the user to create some interactive grpahs with shiny and plotly
}
\examples{
PreAn()
}
