% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{This function, given a set of numerical input, return some basic statistical operation}
\usage{
analysis(x, y)
}
\arguments{
\item{x}{set of numerical input, possibly collected into a data frame}

\item{y}{set of numerical input, possibly collected into a data frame}
}
\value{
mean, variance, standard deviation, maximum value and its position, minimum value
and its position, correlation, simple regression, line plot, bar plot, box plot and scatter plot.
}
\description{
This function, given a set of numerical input, return some basic statistical operation
}
\examples{
To run an example, please, download an example dataset with this command:
data <- data.frame(getDataset()). Then, type and run: analysis(data$Y,data$Italy).
}
