% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{This function allows the user to view the dataset. In order to obtain the
entire datset into your environment type the following command: data <- data.frame(getDataset())}
\usage{
getDataset()
}
\value{
dataset called "data"
}
\description{
This function allows the user to view the dataset. In order to obtain the
entire datset into your environment type the following command: data <- data.frame(getDataset())
}
\examples{
getDataset()
}
