% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{getDataset: this function allows the user to get the dataset}
\usage{
getDataset()
}
\value{
dataset called "data"
}
\description{
getDataset: this function allows the user to get the dataset
}
\examples{
getData()
}
