% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{PreAn}
\alias{PreAn}
\title{ShinyApp: this function plots some interactive grpahs using an example dataset, available
in your environment typing: data <- data.frame(getDataset())}
\usage{
PreAn()
}
\value{
Line plot, bar plot, box plot, scatter plot
}
\description{
ShinyApp: this function plots some interactive grpahs using an example dataset, available
in your environment typing: data <- data.frame(getDataset())
}
\examples{
PreAn()
}
